requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
idna==3.4
instaloader==4.9.5
mysql-connector-python==8.0.31
protobuf==3.20.1
requests==2.28.1
urllib3==1.26.13
wget==3.2
